@@ -11,30 +11,33 @@
     "gh:deploy": "gh-pages -d dist -t true"
   },
   "devDependencies": {
-    "@lehoczky/eslint-config-vue": "^0.0.12",
-    "@nuxt/content": "^2.4.3",
-    "@nuxtjs/tailwindcss": "^6.4.1",
-    "@types/lodash-es": "^4.17.6",
+    "@lehoczky/eslint-config-vue": "^0.0.14",
+    "@nuxt/content": "^2.5.2",
+    "@types/lodash-es": "^4.17.7",
     "@vueuse/core": "^9.13.0",
     "@vueuse/nuxt": "^9.13.0",
-    "eslint": "^8.34.0",
-    "eslint-define-config": "^1.15.0",
+    "eslint": "^8.38.0",
+    "eslint-define-config": "^1.17.0",
     "gh-pages": "^5.0.0",
-    "nuxt": "^3.2.2",
+    "nuxt": "^3.4.0",
     "postcss-html": "^1.5.0",
-    "prettier": "^2.8.4",
-    "prettier-plugin-tailwindcss": "^0.2.3",
-    "stylelint": "^14.16.1",
-    "stylelint-config-prettier": "^9.0.5",
+    "prettier": "^2.8.7",
+    "prettier-plugin-tailwindcss": "^0.2.7",
+    "stylelint": "^15.4.0",
     "stylelint-config-recommended-vue": "^1.4.0",
-    "stylelint-config-standard": "^29.0.0",
+    "stylelint-config-standard": "^32.0.0",
     "stylelint-config-tailwindcss": "^0.0.7",
-    "tailwindcss": "^3.2.7",
+    "tailwindcss": "^3.3.1",
     "tailwindcss-fluid-type": "^2.0.3",
-    "typescript": "^4.9.5"
+    "typescript": "^5.0.4"
   },
   "dependencies": {
     "lodash-es": "^4.17.21"
   },
-  "packageManager": "pnpm@7.14.2"
+  "packageManager": "pnpm@8.2.0",
+  "pnpm": {
+    "overrides": {
+      "consola": "3.0.1"
+    }
+  }
 }
