@@ -1,6 +1,6 @@
 <template>
   <NuxtLink
-    class="transition-color inline-flex w-full items-center py-5 px-4 ease-out hover:bg-base-400 active:scale-95 sm:rounded-md sm:py-2"
+    class="transition-color inline-flex w-full items-center px-4 py-5 ease-out hover:bg-base-400 active:scale-95 sm:rounded-md sm:py-2"
     ><slot
   /></NuxtLink>
 </template>
